Quantity type change

+++ cs3750Fall2013/RecreationOutlet/Warehouse/RecOutletWarehouse/RecOutletWarehouse/Models/MERCHANDISE_TRANSFER.cs
@@ -13,7 +13,7 @@ namespace RecOutletWarehouse.Models
         public string TansferComments { get; set; }
         public short TransferCreatedBy { get; set; }
         public System.DateTime TransferCreatedDate { get; set; }
-        public short Quantity { get; set; }
+        public int Quantity { get; set; }
         public byte QtyTypeID { get; set; }
         public virtual ITEM ITEM { get; set; }
         public virtual QTY_TYPE QTY_TYPE { get; set; }
